Rename files table to task_files
